+++ HISTORY.md
@@ -2,9 +2,10 @@
 
 ## Unreleased
 
-### Public API changes
+### Public API Changes
 * Deprecated WriteOptions::timeout_hint_us. We no longer support write timeout. If you really need this option, talk to us and we might consider returning it.
 * Deprecated purge_redundant_kvs_while_flush option.
+* Removed BackupEngine::NewBackupEngine() and NewReadOnlyBackupEngine() that were deprecated in RocksDB 3.8. Please use BackupEngine::Open() instead.
 
 ## 3.12.0 (7/2/2015)
 ### New Features